requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0
aioresponses==0.7.6
